# swagger_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: French Dictionary
  contact:
    name: Guénolé de Cadoudal
    email: guenoledc@yahoo.fr
  version: 1.0.0
servers:
- url: /french-dictionary/1.0
  description: Default URL
paths:
  /one:
    get:
      description: get one random word
      operationId: one_get
      parameters:
      - name: minLength
        in: query
        description: The min length of the word to be returned. default value is 4
        required: false
        style: form
        explode: true
        schema:
          type: integer
      - name: maxLength
        in: query
        description: The max length of the word to be returned. default value is 100
        required: false
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Word'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /random:
    get:
      description: get one word from the Dictionary choosen randomly
      operationId: random_get
      responses:
        "200":
          description: Successful response, with the following structure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Word'
        "500":
          description: Failed to respond
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
components:
  schemas:
    Error:
      required:
      - message
      type: object
      properties:
        message:
          type: string
        cause:
          type: string
      description: defines an error message
    Word:
      type: object
      properties:
        value:
          type: string
          example: maison
      description: defines one word
      example:
        value: maison